Fix broken horizon-terminal-* slugs
The base16-horizon-terminal-dark theme previously referred to
base16-horizon-dark instead, and similarly for -light. This is
definitely incorrect for the `g:colors_name` variable, as that should
match the name of the colorscheme. I'm not sure whether this was
intentional for base16-shell purposes as I don't use that software so
haven't tested it, but if not then probably the root cause is an error
in whatever .yaml was fed to the template to generate these files (I
looked briefly and couldn't find the source of those).